Moved validating comments to separate rule
