+++ main.go
@@ -19,7 +19,8 @@ func main() {
 				&rules.MSKModuleBackendRule{},
 				&rules.MSKAppTopicsRule{},
 				&rules.MSKTopicNameRule{},
-				rules.NewMSKTopicConfigRule(),
+				&rules.MSKTopicConfigRule{},
+				rules.NewMSKTopicConfigCommentsRule(),
 				&rules.MSKUniqueAppNamesRule{},
 			},
 		},
