+++ rules/msk_topic_config.go
@@ -17,14 +17,6 @@ import (
 // MSKTopicConfigRule checks the configuration for an MSK topic.
 type MSKTopicConfigRule struct {
 	tflint.DefaultRule
-
-	commentsPerFile map[string]hclsyntax.Tokens
-}
-
-func NewMSKTopicConfigRule() *MSKTopicConfigRule {
-	return &MSKTopicConfigRule{
-		commentsPerFile: make(map[string]hclsyntax.Tokens),
-	}
 }
 
 func (r *MSKTopicConfigRule) Name() string {
@@ -105,9 +97,6 @@ func (r *MSKTopicConfigRule) validateTopicConfig(runner tflint.Runner, topic *hc
 		return err
 	}
 
-	if err = r.validateConfigValuesInComments(runner, configKeyToPairMap); err != nil {
-		return err
-	}
 	return nil
 }
 
@@ -655,201 +644,4 @@ func (r *MSKTopicConfigRule) validateRetentionTimeNotDefined(
 		return fmt.Errorf("emitting issue: retention time defined for compacted topic: %w", err)
 	}
 	return nil
-}
-
-type configTimeValueCommentInfo struct {
-	key              string
-	infiniteValue    string
-	baseComment      string
-	issueWhenInvalid bool
-}
-
-var configTimeValueCommentInfos = []configTimeValueCommentInfo{
-	{
-		key:              retentionTimeAttr,
-		infiniteValue:    "-1",
-		baseComment:      "keep data",
-		issueWhenInvalid: false,
-	},
-}
-
-func (r *MSKTopicConfigRule) validateConfigValuesInComments(
-	runner tflint.Runner,
-	configKeyToPairMap map[string]hcl.KeyValuePair,
-) error {
-	for _, configValueInfo := range configTimeValueCommentInfos {
-		if err := r.validateTimeConfigValue(runner, configKeyToPairMap, configValueInfo); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (r *MSKTopicConfigRule) validateTimeConfigValue(
-	runner tflint.Runner,
-	configKeyToPairMap map[string]hcl.KeyValuePair,
-	configValueInfo configTimeValueCommentInfo,
-) error {
-	timePair, hasConfig := configKeyToPairMap[configValueInfo.key]
-	if !hasConfig {
-		return nil
-	}
-
-	msg, err := buildDurationComment(timePair, configValueInfo)
-	if err != nil {
-		return err
-	}
-	if msg == "" {
-		return nil
-	}
-
-	comment, err := r.getExistingComment(runner, timePair)
-	if err != nil {
-		return err
-	}
-
-	if comment == nil {
-		err := runner.EmitIssueWithFix(
-			r,
-			fmt.Sprintf("%s must have a comment with the human readable value: adding it ...", configValueInfo.key),
-			timePair.Key.Range(),
-			func(f tflint.Fixer) error {
-				return f.InsertTextAfter(timePair.Value.Range(), msg)
-			},
-		)
-		if err != nil {
-			return fmt.Errorf("emitting issue: no comment for time value: %w", err)
-		}
-		return nil
-	}
-
-	commentTxt := strings.TrimSpace(string(comment.Bytes))
-	if commentTxt != msg {
-		issueMsg := fmt.Sprintf(
-			"%s value doesn't correspond to the human readable value in the comment: fixing it ...",
-			configValueInfo.key,
-		)
-		err := runner.EmitIssueWithFix(r, issueMsg, comment.Range,
-			func(f tflint.Fixer) error {
-				return f.ReplaceText(comment.Range, msg+"\n")
-			},
-		)
-		if err != nil {
-			return fmt.Errorf("emitting issue: wrong comment for time value: %w", err)
-		}
-	}
-	return nil
-}
-
-func (r *MSKTopicConfigRule) getExistingComment(runner tflint.Runner, pair hcl.KeyValuePair) (*hclsyntax.Token, error) {
-	comments, err := r.getCommentsForFile(runner, pair.Key.Range().Filename)
-	if err != nil {
-		return nil, err
-	}
-
-	// first look for the comment on the same line, after the property definition.
-	// Example: "retention.ms" = "2629800000" # keep data for 30 days
-	afterPropertyIdx := slices.IndexFunc(comments, func(comment hclsyntax.Token) bool {
-		return comment.Range.Start.Line == pair.Key.Range().Start.Line &&
-			comment.Range.Start.Column > pair.Value.Range().End.Column
-	})
-
-	if afterPropertyIdx >= 0 {
-		return &comments[afterPropertyIdx], nil
-	}
-
-	/* second, look for the comment on the previous line, before the property definition. Example:
-	# keep data for 30 days
-	"retention.ms" = "2629800000"
-	*/
-	beforePropertyIdx := slices.IndexFunc(comments, func(comment hclsyntax.Token) bool {
-		return comment.Range.Start.Line == pair.Key.Range().Start.Line-1 &&
-			comment.Range.End.Line == pair.Key.Range().Start.Line
-	})
-	if beforePropertyIdx >= 0 {
-		return &comments[beforePropertyIdx], nil
-	}
-
-	return nil, nil
-}
-
-func (r *MSKTopicConfigRule) getCommentsForFile(runner tflint.Runner, filename string) (hclsyntax.Tokens, error) {
-	cachedComments, hasCachedComments := r.commentsPerFile[filename]
-	if hasCachedComments {
-		return cachedComments, nil
-	}
-	file, err := runner.GetFile(filename)
-	if err != nil {
-		return nil, fmt.Errorf("getting hcl file %s for reading comments: %w", filename, err)
-	}
-
-	tokens, diags := hclsyntax.LexConfig(file.Bytes, filename, hcl.InitialPos)
-	if diags != nil {
-		return nil, diags
-	}
-
-	comments := slices.DeleteFunc(tokens, isNotComment)
-	r.commentsPerFile[filename] = comments
-	return comments, nil
-}
-
-func isNotComment(token hclsyntax.Token) bool {
-	return token.Type != hclsyntax.TokenComment
-}
-
-func buildDurationComment(timePair hcl.KeyValuePair, configValueInfo configTimeValueCommentInfo) (string, error) {
-	var timeVal string
-	diags := gohcl.DecodeExpression(timePair.Value, nil, &timeVal)
-	if diags.HasErrors() {
-		return "", diags
-	}
-
-	if timeVal == configValueInfo.infiniteValue {
-		return fmt.Sprintf("# %s forever", configValueInfo.baseComment), nil
-	}
-
-	timeMillis, err := strconv.Atoi(timeVal)
-	// todo: check what we should do here
-	if err != nil {
-		//nolint:nilerr
-		return "", nil
-	}
-
-	timeUnits, unit := determineTimeUnits(timeMillis)
-
-	msg := fmt.Sprintf("# %s for %d %s", configValueInfo.baseComment, timeUnits, unit)
-	return msg, nil
-}
-
-func determineTimeUnits(millis int) (int, string) {
-	timeInYears := millis / millisInOneYear
-	if timeInYears > 0 {
-		if timeInYears == 1 {
-			return 1, "year"
-		}
-		return timeInYears, "years"
-	}
-
-	timeInMonths := millis / millisInOneMonth
-	if timeInMonths > 0 {
-		if timeInMonths == 1 {
-			return 1, "month"
-		}
-		return timeInMonths, "months"
-	}
-
-	timeInDays := millis / millisInOneDay
-	if timeInDays > 0 {
-		if timeInDays == 1 {
-			return 1, "day"
-		}
-		return timeInDays, "days"
-	}
-
-	timeInHours := millis / millisInOneHour
-	if timeInHours == 1 {
-		return 1, "hour"
-	}
-	return timeInHours, "hours"
 }
