+++ rules/msk_topic_config_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/terraform-linters/tflint-plugin-sdk/helper"
+	"github.com/terraform-linters/tflint-plugin-sdk/tflint"
 )
 
 type topicConfigTestCase struct {
@@ -792,215 +793,6 @@ resource "kafka_topic" "topic_compacted_with_retention_time" {
 	},
 }
 
-var configValueCommentsTests = []topicConfigTestCase{
-	{
-		name: "retention time without comment",
-		input: `
-resource "kafka_topic" "topic_without_retention_comment" {
-  name               = "topic_without_retention_comment"
-  replication_factor = 3
-  config = {
-    "cleanup.policy"   = "delete"
-    "retention.ms"     = "86400000"
-    "compression.type" = "zstd"
-  }
-}`, fixed: `
-resource "kafka_topic" "topic_without_retention_comment" {
-  name               = "topic_without_retention_comment"
-  replication_factor = 3
-  config = {
-    "cleanup.policy"   = "delete"
-    "retention.ms"     = "86400000" # keep data for 1 day
-    "compression.type" = "zstd"
-  }
-}`,
-		expected: []*helper.Issue{
-			{
-				Message: "retention.ms must have a comment with the human readable value: adding it ...",
-				Range: hcl.Range{
-					Filename: fileName,
-					Start:    hcl.Pos{Line: 7, Column: 5},
-					End:      hcl.Pos{Line: 7, Column: 19},
-				},
-			},
-		},
-	},
-	{
-		name: "retention time with wrong comment",
-		input: `
-resource "kafka_topic" "topic_wrong_retention_comment" {
-  name               = "topic_wrong_retention_comment"
-  replication_factor = 3
-  config = {
-    "cleanup.policy"   = "delete"
-    # keep data for 1 day
-    "retention.ms"     = "172800000"
-    "compression.type" = "zstd"
-  }
-}`, fixed: `
-resource "kafka_topic" "topic_wrong_retention_comment" {
-  name               = "topic_wrong_retention_comment"
-  replication_factor = 3
-  config = {
-    "cleanup.policy" = "delete"
-    # keep data for 2 days
-    "retention.ms"     = "172800000"
-    "compression.type" = "zstd"
-  }
-}`,
-		expected: []*helper.Issue{
-			{
-				Message: "retention.ms value doesn't correspond to the human readable value in the comment: fixing it ...",
-				Range: hcl.Range{
-					Filename: fileName,
-					Start:    hcl.Pos{Line: 7, Column: 5},
-					End:      hcl.Pos{Line: 8, Column: 1},
-				},
-			},
-		},
-	},
-	{
-		name: "retention time good infinite comment",
-		input: `
-resource "kafka_topic" "topic_good_retention_comment_infinite" {
-  name               = "topic_good_retention_comment_infinite"
-  replication_factor = 3
-  config = {
-    # keep data in hot storage for 1 day
-    "local.retention.ms"    = "86400000"
-    "remote.storage.enable" = "true"
-    "cleanup.policy"        = "delete"
-    # keep data forever
-    "retention.ms"          = "-1"
-    "compression.type"      = "zstd"
-  }
-}`,
-		expected: []*helper.Issue{},
-	},
-	{
-		name: "retention time in months",
-		input: `
-resource "kafka_topic" "topic_good_retention_comment_months" {
-  name               = "topic_good_retention_comment_months"
-  replication_factor = 3
-  config = {
-    # keep data in hot storage for 1 day
-    "local.retention.ms"    = "86400000"
-    "remote.storage.enable" = "true"
-    "cleanup.policy"        = "delete"
-    "retention.ms"          = "5184000000" # keep data for 2 months 
-    "compression.type"      = "zstd"
-  }
-}`,
-		expected: []*helper.Issue{},
-	},
-	{
-		name: "retention time in years",
-		input: `
-resource "kafka_topic" "topic_good_retention_comment_years" {
-  name               = "topic_good_retention_comment_years"
-  replication_factor = 3
-  config = {
-    # keep data in hot storage for 1 day
-    "local.retention.ms"    = "86400000"
-    "remote.storage.enable" = "true"
-    "cleanup.policy"        = "delete"
-    "retention.ms"          = "31536000000" # keep data for 1 year 
-    "compression.type"      = "zstd"
-  }
-}`,
-		expected: []*helper.Issue{},
-	},
-	{
-		name: "retention time less than a day with good comment",
-		input: `
-resource "kafka_topic" "topic_good_retention_comment_less_than_a_day" {
-  name               = "topic_good_retention_comment_less_than_a_day"
-  replication_factor = 3
-  config = {
-    "cleanup.policy"   = "delete"
-    "retention.ms"     = "21600000" # keep data for 6 hours
-    "compression.type" = "zstd"
-  }
-}`,
-		expected: []*helper.Issue{},
-	},
-	{
-		// checking that multiple insertions won't affect comments positions
-		name: "multiple insertions don't affect comments",
-		input: `
-resource "kafka_topic" "topic_without_policy_and_retention" {
-  name               = "topic_without_policy_and_retention"
-  replication_factor = 3
-  config = {
-    "compression.type" = "zstd"
-  }
-}
-
-resource "kafka_topic" "topic_outdated_retention_comment_years" {
-  name               = "topic_outdated_retention_comment_years"
-  replication_factor = 3
-  config = {
-    # keep data in hot storage for 1 day
-    "local.retention.ms"    = "86400000"
-    "remote.storage.enable" = "true"
-    "cleanup.policy"        = "delete"
-    "retention.ms"          = "63072000000" # keep data for 1 year 
-    "compression.type"      = "zstd"
-  }
-}`,
-		fixed: `
-resource "kafka_topic" "topic_without_policy_and_retention" {
-  name               = "topic_without_policy_and_retention"
-  replication_factor = 3
-  config = {
-    "cleanup.policy"   = "delete"
-    "retention.ms"     = "???"
-    "compression.type" = "zstd"
-  }
-}
-
-resource "kafka_topic" "topic_outdated_retention_comment_years" {
-  name               = "topic_outdated_retention_comment_years"
-  replication_factor = 3
-  config = {
-    # keep data in hot storage for 1 day
-    "local.retention.ms"    = "86400000"
-    "remote.storage.enable" = "true"
-    "cleanup.policy"        = "delete"
-    "retention.ms"          = "63072000000" # keep data for 2 years
-    "compression.type"      = "zstd"
-  }
-}`,
-		expected: []*helper.Issue{
-			{
-				Message: "missing cleanup.policy: using default 'delete'",
-				Range: hcl.Range{
-					Filename: fileName,
-					Start:    hcl.Pos{Line: 5, Column: 3},
-					End:      hcl.Pos{Line: 7, Column: 4},
-				},
-			},
-			{
-				Message: "retention.ms must be defined on a topic with cleanup policy delete",
-				Range: hcl.Range{
-					Filename: fileName,
-					Start:    hcl.Pos{Line: 5, Column: 3},
-					End:      hcl.Pos{Line: 7, Column: 4},
-				},
-			},
-			{
-				Message: "retention.ms value doesn't correspond to the human readable value in the comment: fixing it ...",
-				Range: hcl.Range{
-					Filename: fileName,
-					Start:    hcl.Pos{Line: 18, Column: 45},
-					End:      hcl.Pos{Line: 19, Column: 1},
-				},
-			},
-		},
-	},
-}
-
 var goodConfigTests = []topicConfigTestCase{
 	{
 		name: "good topic definition without tiered storage",
@@ -1063,7 +855,7 @@ func Test_MSKTopicConfigRule(t *testing.T) {
 
 	for _, tc := range allTests {
 		t.Run(tc.name, func(t *testing.T) {
-			rule := NewMSKTopicConfigRule()
+			rule := &MSKTopicConfigRule{}
 			runner := helper.TestRunner(t, map[string]string{fileName: tc.input})
 			require.NoError(t, rule.Check(runner))
 
@@ -1080,7 +872,7 @@ func Test_MSKTopicConfigRule(t *testing.T) {
 	}
 }
 
-func setExpectedRule(expected helper.Issues, rule *MSKTopicConfigRule) {
+func setExpectedRule(expected helper.Issues, rule tflint.Rule) {
 	for _, exp := range expected {
 		exp.Rule = rule
 	}
